requirements.txt
# Core ML Dependencies
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.4
pillow>=10.3.0
scikit-learn>=1.3.0

# Data Processing
pyyaml>=6.0.1
tqdm>=4.66.5

# Web API (Optional)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Camera Support (Optional)
opencv-python>=4.10.0.84

# CLI Tools
typer>=0.12.5

# Visualization
matplotlib>=3.8.0

# OCR Dependencies
pytesseract>=0.3.10
rapidfuzz>=3.0.0

# Development Tools
pytest>=7.4.0

# Large JSON streaming
ijson>=3.3
